README.md
This project created by Kawin Khamkathok simulates a Blackjack game using a single deck of cards and calculates probabilities associated with the game, such as the likelihood to bust or win. To use the program all the user needs to do is click run and then it will show the player's hand and the dealers hand along with the winning and bust probability currently and if the player were to hit and draw and card.